refactored code so that patchgan architecture is mutable to conditioning
