+++ patchgan/__init__.py
@@ -1,8 +1,6 @@
-from .unet import UNet
-from .disc import Discriminator
 from .patchgan import PatchGAN
 from .version import __version__
 
 __all__ = [
-    'UNet', 'Discriminator', 'PatchGAN', '__version__'
+    'PatchGAN', '__version__'
 ]
